Incremental updater retest: * allow inc updater to start new db from scratch * always report localization directly to indexer
git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/branches/lucene-search-2.1@41273 dd0e9695-b195-4be7-bd10-2dea1a65a6b6

+++ src/org/wikimedia/lsearch/frontend/SearchDaemon.java
@@ -373,7 +373,7 @@ public class SearchDaemon extends HttpHandler {
 	 */
 	private void sendResultLine(double score, String namespace, String title) {
 		try{
-		sendOutputLine(score + " " + namespace + " " + encodeTitle(title));
+		sendOutputLine((float)score + " " + namespace + " " + encodeTitle(title));
 		} catch(Exception e){
 			log.error("Error sending result line ("+score + " " + namespace + " " + title +"): "+e.getMessage());
 		}
